+++ conftest.py
@@ -12,6 +12,14 @@
 
 """Pytest dynamic configuration."""
 
+import hypothesis
+
+hypothesis.settings.register_profile(
+    "default",
+    deadline=None,  # Account for slower CI workers
+    print_blob=True,  # Always print code to use with @reproduce_failure
+)
+
 pytest_plugins = [
     "pytest_qiskit_aqt",
 ]