+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 * Add a Grover-based 3-SAT solver example #31
+* Wrap `Rxx` angles to [0, π/2] instead of [-π/2, π/2] #37
 * Wrap single-qubit rotation angles to [0, π] instead of [-π, π]  #39
-* **Breaking change** Remove provider for legacy API #40
+* Remove provider for legacy API #40
 * Automatically load environment variables from `.env` files #42
 
 ## qiskit-aqt-provider v0.9.0